// GARI_Umbria_2026-08-03/GiasNG/src/app/Model/siti.enum.ts
export enum Enum_SiteRedirector {
    GiasLan = 0,
    Sito_AgronicaStampe = 1,
    Sito_AgronicaBio = 2,
    Sito_AgronicaView = 3,
    Sito_PianoConcimazione = 4,
    Sito_AgronicaAnalisi = 5,
    Sito_GiasOnline = 6,
    Sito_AgronicaAudit = 7,
    Sito_AgronicaPlanning = 8,
    Sito_AgronicaPUA = 9,
    Sito_AgronicaManutenzione = 10,
    Sito_AgronicaSicurezzaLavoro = 11,
    Sito_AgronicaProfilazione = 12,
    Sito_AgronicaStampe_NewMode = 13,
    Sito_AgronicaPianiCampionamento = 14,
    Sito_AgronicaAgenda_2010 = 15,
    Sito_AgronicaSincronizzatore = 16,
    Sito_AgronicaAnalisi_2010 = 17,
    Sito_AgronicaGlobalGAP = 18,
    Sito_AgronicaPianiSemina = 19,
    Sito_GiasOnline_2010 = 20,
    Sito_AgronicaMeteo = 21,
    Sito_AgronicaCheckCOOP = 22,
    Ws_GiasPalmWebService_2008 = 23,
    Sito_AgronicaStampe_2010 = 24,
    Sito_AgronicaSementi = 25,
    Sito_AgronicaLabQualita = 26,
    Sito_PianoConcimazione_2017 = 27,
    Sito_AgronicaWebApiProfilatore = 28,
    Sito_AgronicaUma = 29,
    Sito_AgronicaDomandaIrrigua = 30,
    GiasAPP = 100,
    AgroGSB = 200,
    GiasNG = 300,
}


export enum enum_PagineAgenda_2010{
    Menu = 1,
    Pagina_Trattamenti = 2,
    Pagina_Fertilizzazione = 3,
    Pagina_Installazione_Trapppole = 4,
    Pagina_Lavorazioni = 5,
    Pagina_Reinnesco_Trapppole = 6,
    Pagina_Rilievi = 7,
    Pagina_Irrigazione = 8,
    Pagina_RilievoPiogge = 9,
    Pagina_Semina_Trapianto = 10,
    Pagina_Ricette_Lista = 11,
    Pagina_Ricette_Edit = 12,
    Pagina_DocumentoContabileGenerico = 13,
    Pagina_AnalisiCosti = 14,
    Pagina_FormProdotto = 15,
    Pagina_FiltrinoImprese = 16,
    Pagina_GestioneMagazzini = 17,
    Pagina_GestioneContatti = 18,
    Pagina_Distribuzione_Insetti = 19,
    Pagina_MonitoraggioCorpiEstranei = 20,
    Pagina_Operazione_Di_Cura = 21,
    Pagina_Raccolta = 22,
    Pagina_Liquidazione_soci = 23,
    Pagina_Filtrone = 24,
    Pagina_MenuStampe = 25,
    Menu_BS = 26,
    Menu_BS_TrackMode = 27,
    AnalisiDatiCura = 28,
    BI_SchedeRilievi = 29,
    Pagina_GestioneMagazziniBS = 30,
    Pagina_Trattamenti_B = 31,
    Pagina_NonCOnformita_Crea = 32,
    Pagina_Manutenzione_Macchine = 33,
    Pagina_DuplicaOperazione = 34,
    Pagina_Verifica_Conformita = 35,
    Report_Percorsi = 36,
    Pagina_Menu_Liquidazione_soci = 37,
    Pagina_NonConformita_Lista = 38,
    Pagina_Analisi_Progetti = 39,
    Pagina_Analisi_Meteo = 40,
    Pagina_RicetteStampa = 41,
    Pagina_Anagrafica_Ereditatore = 42,
    Pagina_index = 43,
    Pagina_Anagrafica_Appezzamento = 44,
    Pagina_Anagrafica_Campo = 45,
    Pagina_Anagrafica_Impianto = 46,
    Pagina_Anagrafica_Impresa = 47,
    Pagina_Anagrafica_Contatto = 48,
    Pagina_Anagrafica_Fabbricato = 49,
    Pagina_Anagrafica_Macchina = 50,
    Pagina_Anagrafica_Particella = 51,
    Pagina_Anagrafica_Centro = 52,
    Pagina_Scadenzario_Lista = 53,
    Pagina_Scadenzario_CreaModifica = 54,
    Pagina_Anagrafica_Menu = 55,
    Pagina_reportSostenibilita = 56,
    Pagina_Menu_Lavorazioni = 57,
    visite = 58,
    Gis = 59,
    Pagina_Visite_Lista = 60,
    Pagina_Visite_ListaDettagli = 61,
    Pagina_Gestione_Costi = 62,
    FormProdotto = 63,
    Pagina_FatturaElettronica = 64,
    Pagina_ReportVendite = 65,
    Pagina_KendoEditor = 66,
    Pagina_Zoo_Animale = 67,
    Pagina_Zoo_Carico = 68,
    Pagina_Zoo_Configurazione = 69,
    Pagina_Zoo_Control_Panel = 70,
    Pagina_Zoo_Pesatura = 71,
    Pagina_Zoo_Scarico = 72,
    Pagina_Zoo_Spostamento = 73,
    Pagina_AttivitaXCentri_Aziendali = 74,
    Pagina_RicercaDocContabili = 75,
    Pagina_GestioneAllegati = 76,
    DSSWidget = 77,
    Pagina_Anagrafica_Contatto_New = 78,
    Pagina_NonConformita_Anagrafiche = 39,
    Pagina_NonConformita_Impostazioni = 40,
    Pagina_Stime_Produzione = 79,
    Pagina_RicercaTrasferimenti = 80,
    Pagina_TrasferimentoMagazzino = 81,
    Pagina_MVVElettronico = 82,
    Pagina_InvestimentoCatasto = 83,
    Pagina_UMA_Richieste = 84,
    Pagina_UMA_Elenco = 85,
    Pagina_DSS_Difesa = 86,
    Pagina_Analisi_Rilievi = 87,
    Pagina_Documentale_Lista = 88,
    Pagina_Documentale_CreaModifica = 89,
    Pagina_Modifica_Multipla_PianoColturale = 90,
    Pagina_Gestione_Esercizi = 91,
    Pagina_DAA_Elettronico = 92,
    Pagina_DocContabile = 93,
    Pagina_TrattamentiPostRaccolta = 94,
    Pagina_RilieviBS = 95,
    Pagina_IrrigazioneBS = 96,
    Pagina_GestioneRifiuti = 97,
    Pagina_ZooAlimentazione = 98,
    Pagina_ZooAltreLavorazioni = 99,
    Pagina_Verifica_DoseConsigliataBS = 100,
    Pagina_RilievoEffluenti = 101,
    Pagina_Gis_StrumentoDiRipartoCatasto = 102,
    Pagina_Gis_ImportazioneDati = 103,
    Pagina_Gis_EsportazioneDati = 104,
    Pagina_SementiInterferenzeSmall = 105,
    Pagina_Zoo_Trattamento = 106,
    Pagina_Anagrafica_Menu_Zoo = 107,
    Pagina_DatiReteAcqua = 108,

    Pagina_PlugIn_Riepilogo_Meteo = 112,
    Pagina_PlugIn_Monitoraggio = 113,
    Pagina_PlugIn_IndicatoriDSS = 114,
    Pagina_DSS_Irrigazione = 121
}

export enum enum_PagineAnalisi_2010{
    Pagina_Analisi = 1,
Pagina_Gestione_Laboratori = 2,
Stampa_Analisi_Fitofarmaci = 3,
Gestione_Analisi_NEW = 4,
Gestione_Laboratori_NEW = 5
}

export enum enum_PaginePianoConcimazione_2017{
    MenuBS = 0,
    PCB_Inserimento = 1,
    PCB_InserimentoMultiplo = 2,
    PUA_Dichiarazione_Effluenti = 3,
    MenuBS_PUA = 4,
    PUA_Piano_Distribuzione = 5,
    FiltraEdEsporta = 6,
    MenuBs_Piano_Nutrizionale = 7,
    Piano_Nutrizionale = 8
}